Use skb_alloc and ->destructor for TxBuffer management
In existing hardware edrv`s, coherent DMA-able memory is reserved and
allocated via edrv_AllocTxBuffer.c
I don't know why, they chose this over using streaming DMA mappings,
but anyhow, there is no benefit in doing our own buffer allocation here.

The access pattern is favorable too, TxBuffers are allocated at the
start are reused, so no reason not to use kmalloc here.

As we are the network stack, so to say, no one else will mess with
the ->destructor field, so we can use it as Tx confirmation mechanism.
And we can drop that ->is_lock_protected abomination without having to
change any code whatsoever in the DLL.

This is untested and still wrong: We transmit with interrupts disabled,
which isn't required to work. Use netpoll or transmission from tasklet
instead.